requirements.txt
Flask==3.0.2              # Pour créer une API locale ou un serveur web si nécessaire
requests==2.31.0      # Pour effectuer des requêtes HTTP à l'API OpenAI Vision
#PyBluez==0.23             # Pour la communication Bluetooth (sur des systèmes compatibles)
opencv-python-headless==4.9.0.80  # Pour la capture et le traitement d'images
gtts==2.5.1            # Google Text-to-Speech, pour la synthèse vocale
pyttsx3==2.90             # Une alternative à gTTS qui fonctionne hors ligne
numpy==1.26.1             # Requis par OpenCV pour le traitement d'images
Pillow==10.2.0             # Pour le traitement d'images, peut-être utile selon vos besoins
python-dotenv==0.20.0  # Pour charger des variables d'environnement depuis un fichier .env
aiohttp==3.9.3            # Pour effectuer des requêtes HTTP asynchrones
ultralytics==8.1.27
bleak==0.21.1             # Pour la communication Bluetooth Low Energy (BLE)
opencv-python~=4.9.0.80
pynput~=1.7.6
torch==2.2.1
torchvision==0.17.1
pip~=24.0
attrs~=23.2.0
wheel~=0.43.0
Jinja2~=3.1.3
setuptools~=69.1.1
Cython~=3.0.9
pandas~=2.2.1
pygame==2.5.2
numpy==1.26.1
pymysql==1.1.0